+++ examples/empirical/implementation/bai_ng_utils.py
